refactored a good portion of meals ajax into response
