+++ lib/response-handlers/meal-responses.js
@@ -1,8 +1,6 @@
 const $ = require('jquery')
 const API = "https://rocky-earth-59921.herokuapp.com";
-// const displayDiary = require("../ajax_requests/meal")
-// const handleError = require("../ajax_requests/meal")
-import {displayDiary, handleError} from '../ajax_requests/meal'
+import {displayDiary, handleError, totalCalories} from '../ajax_requests/meal'
 
 const goalKey = {
   Breakfast: 400,
@@ -11,7 +9,7 @@ const goalKey = {
   Snack: 200
 }
 
-function mainMealResponse(data) {
+export function mainMealResponse(data) {
   data.forEach(function(meal) {
     var sum = 0
     meal.foods.forEach(function(food) {
@@ -28,7 +26,7 @@ function mainMealResponse(data) {
 
 
 
-function mealFoodListener(foodId, meal) {
+export function mealFoodListener(foodId, meal) {
   $(`#${meal.name.toLowerCase()} [data-id=${foodId}] .delete-icon`).on('click', function(event) {
     var item = $(event.target)
     $.ajax({
@@ -43,4 +41,68 @@ function mealFoodListener(foodId, meal) {
   })
 }
 
-module.exports = {mainMealResponse, mealFoodListener}
+export function totalsTable() {
+  $('#totals').append(
+    `<tr>
+      <td>Goal Calories </td>
+       <td>2000 </td> </tr>
+    <tr>
+      <td>Calories Consumed </td>
+      <td> ${allMealTotalCals()}</td>
+      </tr>
+    <tr>
+    <td>Remaining Calories </td>
+      <td class="remaining-totals"> ${2000 - allMealTotalCals()}  </td>
+    </tr>
+    `
+  )
+  renderRemainingColor();
+}
+
+function allMealTotalCals () {
+  var allCals = totalCalories();
+  return allCals
+
+}
+
+export function renderRemainingColor() {
+    $(`td.remaining-totals:contains('-')`).addClass('red').removeClass('green')
+    $(`td.remaining-totals:not(:contains('-'))`).addClass('green').removeClass('red')
+}
+
+export function renderFoods(foods) {
+  $('#foods-table-headers').append("<h2 class='text-center'>Foods</h2><br>"
+                            + "<p>Add selected to:</p>")
+  $('#foods-table').append("<table class='table'>"
+                    + "<thead><tr><th></th><th>Name</th><th>Calories</th></tr></thead><tbody>"
+                    + createFoodRows(foods)
+                    + "</tbody></table>")
+}
+
+function createFoodRows(foods) {
+  var rows = "";
+  $.each(foods, function(index, food) {
+    rows += `<tr><td><input type='checkbox' name='food' value='food' foodId=${food["id"]}></td><td>${food["name"]}</td><td>${food["calories"]}</td></tr>`
+  })
+  return rows
+}
+
+export function displayAddButtons(meals) {
+  $.each(meals, function(index, meal) {
+    $('#add-selected').append(`<button type='button' name='button' class='btn btn-primary' mealId="${meal["id"]}">
+    ${meal["name"]}</button>`)
+  })
+}
+
+export function clearElements () {
+  $('.meal-table').children('tbody').empty()
+  $('#totals').children().html('')
+  $('#add-selected').children().remove();
+  $('tfoot').empty()
+}
+
+
+export function clearFoods() {
+  $('#foods-table').html('')
+  $('#foods-table-headers').html('')
+}
